// src/app/Models/ForAllScheduleDto.ts
export class forAllSchedule {
  buId: string;
  buName: string;
  candidateId: string;
  candidateName: string;
  candidateType: string;
  contactNumber: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  customerName: string;
  duration: string;
  employmentType: string;
  experience: string;
  id: string;
  interviewerId: string;
  interviewerName: string;
  jrId: string;
  jrNumber: string;
  location: string;
  message: string;
  mode: string;
  noOfOpenings: string;
  placementFor: string;
  recStatus: string;
  resultRemarks: string;
  roleName: string;
  round: string;
  scheduleDateTime: string;
  scheduleRemarks: string;
  sex: string;
  status: string;
  timeZone: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  venue: string;
}